Update Nc_conv usage
Nc_conv should only act to truncate the full SHC data from the GPUMD run. If Nc_conv is provided and is valid, it should only update Nc. If it is not provided, then we use the full set of data from the SHC run which will have Nc correlation points. This commit makes sure Nc_conv is only used where appropriate.